Move AnimationLevel to _types.
See review comment: https://github.com/Textualize/textual/pull/4062#discussion_r1483111016.
