+++ src/textual/_types.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any, Awaitable, Callable, List, Union
+from typing import TYPE_CHECKING, Any, Awaitable, Callable, List, Literal, Union
 
 from typing_extensions import Protocol
 
@@ -40,3 +40,6 @@ WatchCallbackType = Union[
     Callable[[Any, Any], None],
 ]
 """Type used for callbacks passed to the `watch` method of widgets."""
+
+AnimationLevel = Literal["none", "basic", "full"]
+"""The levels that the [`TEXTUAL_ANIMATIONS`][textual.constants.TEXTUAL_ANIMATIONS] env var can be set to."""
